skinscan-backend/requirements.txt
# Core Django
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1

# Database
psycopg2-binary==2.9.9

# Security
django-ratelimit==4.1.0

# Authentication & Security
PyJWT==2.8.0
python-decouple==3.8
bcrypt==4.1.2

# Image Processing & AI
Pillow==10.1.0
opencv-python-headless==4.8.1.78
tensorflow==2.15.0
numpy==1.24.3

# Google Cloud (for production)
google-cloud-storage==2.10.0
